Unified Storage: Use dialect Ident for ANALYZE and reuse bulk iterator helper
Quote the resource_history table via the dialect's Ident instead of a hardcoded
literal, and move the bulk integration test into backend_bulk_test.go, reusing an
exported ToBulkIterator helper rather than a bespoke iterator.
